Improve connection robustness when reconnecting.
This commit improves the connection robustness on reconnect attempts.
Any ZK work on re-connect is now down asynchronously in order to not
interfere with the gate event thread.

Additionally, the event processing thread does not process any events
when the service is disconnected. Failed events will be put back into
the event queue for later processing as long as the service is not
closed.

Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>