Bug 360503 - alternate storage for job history
This is the first part of a rework which extracts the job history
storage into a customizable SPI. A default implementation is provided
based on cloud preferences.

This commit also started to separate the job storage into a separate
interface that may offer an additional storage SPI for job and schedule
configuration data.

Bug: 360503
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>