Separate state for context working copies.
It turns out that shared objects between an original context and a
working copy put some real limitations on the table. For example, it's
not possible to just override a dependency of a context object. This
commit removes the shared state and creates an individual object tree
for local contexts.

Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>