Support providing jobs using a scanning and annotation based approach.
Similar to other areas, this change add support for providing jobs using
an annotation based approach. All clients need to do is providing a
service component using the ScanningJobProviderComponent in this bundle.
It will then scan the bundle for job classes and use the job types
discovered via the JobType annotation.

Change-Id: Ifbe2e3f56bdcd7157055e19d88f1e34df0686cc8
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>