Move additional modules into profile which can be deactivated.
Sometimes we only want to do a quick build to verify that everything
compiles and that tests are running without producing the full product
packages. This change moves those modules into a profile which is active
by default (for backwards compatibility reasons and to simplify building
on the command line). However, on Hudson it can be deactivated to allow
Gerrit triggered builds finish early.

Change-Id: Id1d49e435b0c16a5fc8411a7df4a4f2e8bf37bd5
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>