Bug 432561: API for a distributed event bus
This commit adds API for a distributed event bus. The bus is capable of
working with any event type and specific transports. Transports are
provided as a cloud service. They represent are a low level of working
with distributed events. The event bus adds a layer on top of that which
allows working with custom event types and listeners.

Change-Id: I12d347b53f465d82c71785f8c5df7a6bacb2628c
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=432561