Rework Logback configuration.
There are a few open ends within the Logback configuration area
addressed by this change. The most significant is a change in the model
to avoid mistakes done the road (by make appenders a loggers not a map
that can be accessed freely).

Also this changes implements proper loading and saving of common
appender attributes (such as name an threshold).

Change-Id: I4d8c3e4cc6d0fea56f8e8ad31f029244a7d1890b
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>